requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
aiosqlite==0.20.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
itsdangerous==2.2.0
jinja2==3.1.4
email-validator==2.2.0
pillow==11.0.0
python-dotenv==1.0.1
